accessible/tests/mochitest/table/test_struct_listbox.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Table accessible tree and table interface tests for XUL listboxes">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../table.js"></script>

  <script type="application/javascript">
  <![CDATA[
    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // Multicolumn listbox.

      var cellsArray = [
        [kDataCell, kDataCell],
        [kDataCell, kDataCell]
      ];

      testTableStruct("listbox1", cellsArray);

      //////////////////////////////////////////////////////////////////////////
      // Multicolumn listbox with header.

      var cellsArray = [
        [kDataCell, kDataCell, kDataCell],
        [kDataCell, kDataCell, kDataCell],
        [kDataCell, kDataCell, kDataCell]
      ];

      testTableStruct("listbox2", cellsArray, kListboxColumnHeader);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox style="overflow: auto;">
  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=512424"
       title="implement IAccessibleTable2">
      Mozilla Bug 512424
    </a>

    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <vbox flex="1">

    <label control="listbox1" value="multicolumn listbox: "/>
    <listbox id="listbox1">
      <listcols>
        <listcol flex="1"/>
        <listcol flex="1"/>
      </listcols>
      <listitem>
        <listcell label="cell1"/>
        <listcell label="cell2"/>
      </listitem>
      <listitem>
        <listcell label="cell1"/>
        <listcell label="cell2"/>
      </listitem>
    </listbox>

    <label control="listbox2" value="multicolumn listbox with header"/>
    <listbox id="listbox2">
      <listhead>
        <listheader label="header1"/>
        <listheader label="header2"/>
        <listheader label="header3"/>
      </listhead>
      <listcols>
        <listcol flex="1"/>
        <listcol flex="1"/>
        <listcol flex="1"/>
      </listcols>
      <listitem>
        <listcell label="cell0"/>
        <listcell label="cell1"/>
        <listcell label="cell2"/>
      </listitem>
      <listitem>
        <listcell label="cell3"/>
        <listcell label="cell4"/>
        <listcell label="cell5"/>
      </listitem>
      <listitem>
        <listcell label="cell6"/>
        <listcell label="cell7"/>
        <listcell label="cell8"/>
      </listitem>
    </listbox>
  </vbox>
  </hbox>

</window>
